fix static query limit
